Rename folder email to email-and-password.
