+++ packages/authentication/src/email-and-password/email-and-password.schema.spec.ts
