+++ packages/authentication/src/email-and-password/email-authenticator.service.ts
