+++ packages/authentication/src/email-and-password/index.ts
