+++ packages/authentication/src/email-and-password/parse-password.parser.spec.ts
