+++ packages/authentication/src/email-and-password/parse-password.parser.ts
