+++ packages/authentication/src/email-and-password/validate-email-credentials-format.pre-hook.ts
