+++ packages/authentication/src/index.ts
@@ -1,4 +1,4 @@
-export * from './email';
+export * from './email-and-password';
 export * from './authorization';
 export * from './authenticate.pre-hook';
 export * from './authentication.controller-factory';
